+++ src/main/java/ru/sber/internship/repository/ProductRepository.java
@@ -5,7 +5,10 @@ import org.springframework.stereotype.Repository;
 import ru.sber.internship.entity.Client;
 import ru.sber.internship.entity.Product;
 
+import java.util.List;
+
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-    Product findById(int id);
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    Product findById(long id);
+
 }
